refactor(gptmail): consolidate scripts/email into packages/gptmail
- Add export_to_maildir and import_from_maildir methods to lib.py
- Add export_maildir, import_maildir, and check_complexity CLI commands
- Remove scripts/email/ directory (gptmail is now canonical)
- Update references in README, tests, and documentation
- Fix type annotation in tasks/cli.py (pre-existing issue)

Closes: gptme/gptme-contrib#166
