--- scripts/email/pyproject.toml
@@ -1,36 +1,0 @@
-[project]
-name = "gptme-email"
-version = "0.1.0"
-description = "Email automation for gptme agents with shared communication utilities"
-requires-python = ">=3.10"
-dependencies = [
-    "requests",
-    "python-dateutil",
-    "markdown",
-    "flask>=2.0.0",
-    "werkzeug>=2.0.0",
-]
-
-[project.optional-dependencies]
-dev = [
-    "pytest>=7.0.0",
-    "mypy>=1.0.0",
-    "ruff>=0.1.0",
-]
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.wheel]
-packages = ["communication_utils"]
-
-[tool.ruff]
-line-length = 100
-target-version = "py310"
-
-[tool.mypy]
-python_version = "3.10"
-warn_return_any = true
-warn_unused_configs = true
-disallow_untyped_defs = true
